fix: profile connected account display
Elements display the "Connected accounts" heading even if
all providers (of any type) are disabled. This is because
the disconnect account action just adds all a user model's
identities to allowed values, regardless of their enabled
status. This makes the action enabled (instead of suspended
when all providers are disabled) and hence elements
displays the heading.

This changes that by using the previously established function
for checking whether an identity is related to a disabled
provier. This also moves it to a utility function because
it is used in more than one place. It is placed in the
util package because I did not want to create a dependency
of the models package on the config package.
