+++ backend/flow_api/flow/profile/action_disconnect_thirdparty_oauth_provider.go
@@ -9,6 +9,7 @@ import (
 	"github.com/teamhanko/hanko/backend/v3/flow_api/flow/shared"
 	"github.com/teamhanko/hanko/backend/v3/flowpilot"
 	"github.com/teamhanko/hanko/backend/v3/persistence/models"
+	"github.com/teamhanko/hanko/backend/v3/utils"
 )
 
 type DisconnectThirdpartyOauthProvider struct {
@@ -37,8 +38,32 @@ func (a DisconnectThirdpartyOauthProvider) Initialize(c flowpilot.Initialization
 
 	input := flowpilot.StringInput("identity_id").Required(true)
 
+	type allowed struct {
+		ProviderID string
+		IdentityID uuid.UUID
+	}
+
+	var allowedValues []allowed
 	for _, identity := range userModel.Identities {
-		input.AllowedValue(identity.ProviderID, identity.ID)
+		if !utils.IsIdentityForDisabledProvider(
+			identity,
+			a.GetDeps(c).Cfg.TenantConfig.ThirdParty.Providers.GetEnabled(),
+			a.GetDeps(c).Cfg.TenantConfig.ThirdParty.CustomProviders.GetEnabled(),
+		) {
+			allowedValues = append(allowedValues, allowed{
+				ProviderID: identity.ProviderID,
+				IdentityID: identity.ID,
+			})
+		}
+	}
+
+	if len(allowedValues) == 0 {
+		c.SuspendAction()
+		return
+	}
+
+	for _, allowedValue := range allowedValues {
+		input.AllowedValue(allowedValue.ProviderID, allowedValue.IdentityID)
 	}
 }
